Configure git to don't use unauthenticated protocol in docker
The `relay` package contains a dependency to `@summa-tx/relay-sol@2.0.2`
package, which downloads one of its sub-dependencies via unathenticated
`git://` protocol. That protocol is no longer supported by GitHub. This
means that in certain situations installation of the package or update
of its dependencies using NPM may result in `The unauthenticated git
protocol on port 9418 is no longer supported` error.
That what was happening as a result of running `go generate ./...`
command during building of Docker image.
As a workaround, we've added a step which configures Git to use
`https://` protocol instead of `git://` when downloading files.